+++ app/Filament/Resources/CloakRuleResource/Pages/CreateCloakRule.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\CloakRuleResource\Pages;
+
+use App\Filament\Resources\CloakRuleResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateCloakRule extends CreateRecord
+{
+    protected static string $resource = CloakRuleResource::class;
+}