+++ app/Filament/Resources/CloakRuleResource/Pages/EditCloakRule.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\CloakRuleResource\Pages;
+
+use App\Filament\Resources\CloakRuleResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditCloakRule extends EditRecord
+{
+    protected static string $resource = CloakRuleResource::class;
+}