+++ app/Filament/Resources/PixelResource/Pages/EditPixel.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\PixelResource\Pages;
+
+use App\Filament\Resources\PixelResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditPixel extends EditRecord
+{
+    protected static string $resource = PixelResource::class;
+}